chore: 支払日選択フィールドを作成し、DatePickerコンポーネントをシンプルにした
Why: ラベルやメッセージはフォーム専用のため
