+++ apps/web/src/features/payments/createPayment/PaymentDateField/index.ts
@@ -1,0 +1,1 @@
+export { PaymentDateField } from "./PaymentDateField"
